Move ParameterSet from template.h to parameter_set.h
