+++ pycode/memilio-simulation/memilio/simulation/utils/parameter_set.h
@@ -1,0 +1,31 @@
+#ifndef PYMIO_PARAMETER_SET_H
+#define PYMIO_PARAMETER_SET_H
+
+#include "memilio/utils/parameter_set.h"
+
+#include "pybind11/pybind11.h"
+
+namespace pymio
+{
+
+template <class ParameterSet>
+auto bind_ParameterSet(pybind11::module& m, std::string const& name)
+{
+    pybind11::class_<ParameterSet> c(m, name.c_str());
+    mio::foreach_tag<ParameterSet>([&c](auto t) {
+        using Tag = decltype(t);
+
+        //CAUTION: This requires ParameterTag::name() to be unique within the ParameterSet
+        c.def_property(
+            Tag::name().c_str(), [](const ParameterSet& self) -> auto& { return self.template get<Tag>(); },
+            [](ParameterSet& self, typename Tag::Type const& v) {
+                self.template get<Tag>() = v;
+            },
+            pybind11::return_value_policy::reference_internal);
+    });
+    return c;
+}
+
+} // namespace pymio
+
+#endif //PYMIO_PARAMETER_SET_H
